Added handles to the actors and runtimes.
