+++ aktoro-raw/src/actor/handle.rs
@@ -1,0 +1,2 @@
+/// TODO: documentation
+pub trait Handle: Unpin + Clone + Send {}
