+++ aktoro-raw/src/lib.rs
@@ -1,10 +1,13 @@
-/// TODO: documentation
+/// TODO: crate documentation
 /// TODO(struct): actor supervisor
 /// TODO(struct): actor schduler
 
 mod context;
 
+/// TODO: documentation
 pub mod actor;
+
+/// TODO: documentation
 pub mod runtime;
 
 pub use actor::Actor;
